+++ app/assets/javascripts/tolk/libraries.js
@@ -1,1 +1,0 @@
-//= require tolk/jquery-1.9.1.min.js