+++ app/controllers/tolk/locales_controller.rb
@@ -1,13 +1,32 @@
 module Tolk
   class LocalesController < Tolk::ApplicationController
-    before_action :find_locale, only: [:show, :update, :incomplete_translations, :completed_translations, :updated_translations]
+    before_action :find_locale, only: [:edit, :update]
 
     def index
       @locales = Tolk::Locale.all.sort_by(&:language_name)
     end
 
+    def create
+      Tolk::Locale.create!(params.require(:tolk_locale).permit(:name))
+
+      redirect_to action: :index
+    end
+    
     def show
-      redirect_to action: :incomplete_translations, id: @locale.name
+      redirect_to action: :edit, id: @locale.name
+    end
+
+    def edit
+      respond_to do |format|
+        format.html do
+          get_phrases
+        end
+
+        format.yaml do
+          data = @locale.to_hash
+          render plain: Tolk::YAML.dump(data)
+        end
+      end
     end
 
     def update
@@ -15,59 +34,27 @@ module Tolk
 
       if @locale.update(translation_params)
         flash[:notice] = "Save Successful!"
-        redirect_back(fallback_location: locale_path(@locale))
+
+        redirect_back(fallback_location: edit_locale_path(@locale))
       else
         flash.now[:alert] = "Not Saved!"
 
-        if params[:form_name] == "completed_translations"
-          @phrases = @locale.phrases_with_translation(params[pagination_param])
-          render "tolk/locales/#{params[:form_name]}"
-        else
-          @phrases = @locale.phrases_without_translation(params[pagination_param])
-          render "tolk/locales/#{params[:form_name]}"
-        end
+        get_phrases
+
+        render "tolk/locales/edit"
       end
-    end
-
-    def incomplete_translations
-      respond_to do |format|
-        format.html do
-          @action_name = params[:action]
-            
-          @phrases = @locale.phrases_without_translation(params[pagination_param])
-        end
-
-        format.atom { @phrases = @locale.phrases_without_translation(params[pagination_param]).per(50) }
-
-        format.yaml do
-          data = @locale.to_hash
-          render plain: Tolk::YAML.dump(data)
-        end
-
-      end
-    end
-
-    def completed_translations
-      @action_name = params[:action]
-
-      @phrases = @locale.phrases_with_translation(params[pagination_param])
-    end
-
-    def updated_translations
-      @phrases = @locale.phrases_with_updated_translation(params[pagination_param])
-      render :all
-    end
-
-    def create
-      Tolk::Locale.create!(params.require(:tolk_locale).permit(:name))
-      redirect_to action: :index
     end
 
     def dump_all
       Tolk::Locale.dump_all
+
       I18n.reload!
-      I18n::JS.export if defined? I18n::JS
-      redirect_to request.referrer
+
+      if defined? I18n::JS
+        I18n::JS.export 
+      end
+
+      redirect_back(fallback_location: locales_path)
     end
 
     def stats
@@ -76,11 +63,14 @@ module Tolk
       respond_to do |format|
         format.json do
           stats = @locales.collect do |locale|
-            [locale.name, {
-              missing: locale.count_phrases_without_translation,
-              updated: locale.count_phrases_with_updated_translation,
-              updated_at: locale.updated_at
-            }]
+            [
+              locale.name, 
+              {
+                missing: locale.count_phrases_without_translation,
+                updated: locale.count_phrases_with_updated_translation,
+                updated_at: locale.updated_at
+              }
+            ]
           end
           render json: Hash[stats]
         end
@@ -90,7 +80,26 @@ module Tolk
     private
 
     def find_locale
-      @locale = Tolk::Locale.where('UPPER(name) = UPPER(?)', params[:id]).first!
+      @locale = Tolk::Locale.find_by!('UPPER(name) = UPPER(?)', params[:id])
+    end
+
+    def get_phrases
+      case params[:filter]
+      when "incomplete"
+        @phrases = @locale.phrases_without_translation
+      when "completed"
+        @phrases = @locale.phrases_with_translation
+      when "updated"
+        @phrases = @locale.phrases_with_updated_translation
+      else
+        @phrases = @locale.phrases.includes(:translations)
+      end
+
+      if params[:q].present? || params[:k].present?
+        @phrases = @phrases.search_phrases(params[:lang], params[:q], params[:k])
+      end
+
+      @phrases = @phrases.send(pagination_method, params[:page])
     end
 
   end