+++ config/routes.rb
@@ -1,12 +1,5 @@
 Tolk::Engine.routes.draw do
-  resources :locales, only: [:index, :show, :update] do
-    member do
-      get :incomplete_translations
-      get :completed_translations
-    end
-  end
-  
-  resource :search
+  resources :locales, only: [:index, :show, :edit, :update]
 
   post "/dump_all" => "locales#dump_all", as: :dump_all_locales
   get "/stats" => "locales#stats"