+++ lib/tolk/sync.rb
@@ -1,5 +1,6 @@
 module Tolk
   module Sync
+
     def self.included(base)
       base.send :extend, ClassMethods
     end
@@ -11,33 +12,41 @@ module Tolk
 
       def load_translations
         if Tolk.config.exclude_gems_token
-          # bypass default init_translations
-          I18n.backend.reload! if I18n.backend.initialized?
+          if I18n.backend.initialized?
+            # bypass default init_translations
+            I18n.backend.reload! 
+          end
+
           I18n.backend.instance_variable_set(:@initialized, true)
+
           translations_files = Dir[Rails.root.join('config', 'locales', "*.{rb,yml}")]
 
           if Tolk.config.block_xxx_en_yml_locale_files
-            locale_block_filter = Proc.new {
-              |l| ['.', '..'].include?(l) ||
+            translations_files =  translations_files.reject{|l| 
+              ['.', '..'].include?(l) ||
                 !l.ends_with?('.yml') ||
                 l.split("/").last.match(/(.*\.){2,}/) # reject files of type xxx.en.yml
             }
-            translations_files =  translations_files.reject(&locale_block_filter)
           end
 
           I18n.backend.load_translations(translations_files)
         else
           I18n.backend.send :init_translations unless I18n.backend.initialized? # force load
         end
+
         translations = flat_hash(I18n.backend.send(:translations)[primary_locale.name.to_sym])
+
         filtered = filter_out_i18n_keys(translations.merge(read_primary_locale_file))
+
         filter_out_ignored_keys(filtered)
       end
 
       def read_primary_locale_file
         primary_file = "#{self.locales_config_path}/#{self.primary_locale_name}.yml"
+
         if File.exist?(primary_file)
-          flat_hash(Tolk::YAML.load_file(primary_file)[self.primary_locale_name])
+          hash = Tolk::YAML.load_file(primary_file)[self.primary_locale_name]
+          flat_hash(hash)
         else
           {}
         end
@@ -66,14 +75,19 @@ module Tolk
         Tolk::Phrase.where.not(key: translations.keys).destroy_all
 
         phrases_by_key = Tolk::Phrase.all.index_by(&:key)
+
         primary_translation_by_phrase_id = primary_locale.translations.index_by(&:phrase_id)
 
         translations.each do |key, value|
-          next if value.is_a?(Proc)
+          if value.is_a?(Proc)
+            next
+          end
+
           # Create phrase and primary translation if missing
           phrase = phrases_by_key[key] || Tolk::Phrase.create!(key: key)
-          translation = primary_translation_by_phrase_id[phrase.id]
-          translation ||= primary_locale.translations.build(phrase: phrase)
+
+          translation = primary_translation_by_phrase_id[phrase.id] || primary_locale.translations.build(phrase: phrase)
+
           translation.text = value
 
           if translation.changed? && !translation.new_record?
@@ -82,7 +96,10 @@ module Tolk
           end
 
           translation.primary = true
-          translation.save! if translation.changed?
+
+          if translation.changed?
+            translation.save!
+          end
         end
       end