+++ test/test_helper.rb
@@ -7,7 +7,6 @@ require "capybara/rails"
 require "mocha/minitest"
 require 'capybara/poltergeist'
 
-
 Rails.backtrace_cleaner.remove_silencers!
 
 # Load support files
@@ -17,8 +16,12 @@ Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 ActiveSupport::TestCase.fixture_path = File.expand_path("../fixtures", __FILE__)
 
 # Configure capybara for integration testing
-Capybara.default_driver   = :poltergeist
-options = {js_errors: false}
-Capybara.register_driver :poltergeist do |app|
-  Capybara::Poltergeist::Driver.new(app, options)
-end
+# Capybara.default_driver   = :poltergeist
+
+# options = {js_errors: false}
+
+# Capybara.register_driver :poltergeist do |app|
+#   Capybara::Poltergeist::Driver.new(app, options)
+# end
+
+Capybara.server = :webrick